added info and small changes in create_dataset.py

+++ create_dataset.py
@@ -4,17 +4,26 @@ import zipfile
 
 
 USE_RGB = False
+DELETE_TEMP = True
 OCCLUSION_BOUNDS = (1, 20)
 
 abs_path = os.path.abspath("./")
 input_dataset_path = f"{abs_path}/input_dataset"
-dataset_path = f"{abs_path}/dataset"
+dataset_path = f"{abs_path}/dataset/{OCCLUSION_BOUNDS[0]}_{OCCLUSION_BOUNDS[1]}"
 temp_path = f"{abs_path}/temp"
 
+print(f"Started with occlusion bounds set as {OCCLUSION_BOUNDS}.")
+
 try:
-    shutil.rmtree(f"{abs_path}/dataset")
+    shutil.rmtree(dataset_path)
 except FileNotFoundError:
     pass
+
+if DELETE_TEMP:
+    try:
+        shutil.rmtree(f"{abs_path}/temp")
+    except FileNotFoundError:
+        pass
 
 os.makedirs(f"{dataset_path}/input", exist_ok=True)
 os.makedirs(f"{dataset_path}/ground_truth", exist_ok=True)
@@ -22,13 +31,24 @@ os.makedirs(f"{temp_path}", exist_ok=True)
 
 files = [f for f in os.listdir(input_dataset_path) if os.path.isfile(os.path.join(input_dataset_path, f))]
 
-for file in files:
+for i, file in enumerate(files):
     with zipfile.ZipFile(f"{input_dataset_path}/{file}", 'r') as zip_ref:
+        print(f"[{i + 1}/{len(files)}] unzipping {file}.")
         zip_ref.extractall(f"{temp_path}")
 
 last_index = 0
 global_index = 0
+all_files = 1
 for root, dirs, files in os.walk(temp_path):
+    if USE_RGB:
+        print(f"{global_index} pairs ready. \t"
+              f"{int(round(all_files / 2, 0))} pairs in current directory to check. \t"
+              f"{int(round(100 * global_index * 4 / all_files, 0))}% of files are used.")
+    else:
+        print(f"{global_index} pairs ready. \t"
+              f"{int(round(all_files / 2, 0))} pairs in current directory to check. \t"
+              f"{int(round(100 * global_index * 2 / all_files, 0))}% of files are used.")
+
     for file in files:
         file_path = os.path.join(root, file)
         file_info = file.rstrip(".png").split("_")
@@ -60,7 +80,13 @@ for root, dirs, files in os.walk(temp_path):
             global_index += 1
             last_index = file_index
 
-try:
-    shutil.rmtree(f"{abs_path}/temp")
-except FileNotFoundError:
-    pass
+    all_files += len(files)
+
+if DELETE_TEMP:
+    print("Deleting temporary directory.")
+    try:
+        shutil.rmtree(f"{abs_path}/temp")
+    except FileNotFoundError:
+        pass
+
+print(f"Created {global_index} pairs of images. Total number of pairs is {int(round(all_files / 2, 0))}.")
